pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Student_Management_System"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.9"
dependencies = [ 
    "mongita"
]  # runtime deps go here

[project.optional-dependencies]
dev = [
    "pre-commit",
    "black",
    "isort",
    "customtkinter",
    "pymongo",
    "python-dotenv"
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]
skip-string-normalization = true
exclude = '''
(
  /(
      \.git
    | \.venv
    | build
    | dist
    | __pycache__
    | tests
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
skip = ["tests", ".venv", "build", "dist", "__pycache__"]
